Refactor project structure: reorganized and split files
